detection+recognition/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.2
opencv-python==4.4.0.44
pdf2image==1.14.0
Pillow==8.0.0
requests==2.24.0
scipy==1.5.2
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
